+++ Windows/TrainingWindow.cs
@@ -52,6 +52,13 @@ public sealed class TrainingWindow : Window
                 ImGui.EndTabItem();
             }
 
+            if (ImGui.BeginTabItem("Recommended"))
+            {
+                ImGui.Spacing();
+                RecommendationsTab.Draw(this.trainingService, this.configuration.Training);
+                ImGui.EndTabItem();
+            }
+
             if (ImGui.BeginTabItem("Lessons"))
             {
                 ImGui.Spacing();